Refactor probe specs into their own sub-tree so that (planned feature) record & replay can depend on them.
